// enums/http_responses.ts
enum httpResponses {
  status200 = "ok",
  status201 = "created",
  status202 = "accepted",
  status400 = "Bad Request",
  status401 = "Unauthorized",
  status402 = "Payment Required",
  status403 = "Forbidden",
  status404 = "Not Found",
  status500 = "Internal Server Error",
  status501 = "Not Implemented",
  status502 = "Bad Gateway",
  status503 = "Service Unavailable",
}

export { httpResponses };